Move to transaction system, bugfixes.

+++ src/main/java/tesseract/api/gt/GTHolder.java
@@ -11,7 +11,7 @@ public class GTHolder {
      * @param cable    The cable connector.
      * @param amperage The initial amperage.
      */
-    protected static long create(IGTCable cable, int amperage) {
+    protected static long create(IGTCable cable, long amperage) {
         return (long) cable.getAmps() << 32 | amperage;
     }
 
@@ -21,7 +21,7 @@ public class GTHolder {
      * @param holder   The long with the packed holder.
      * @param amperage The added amperage.
      */
-    protected static long add(long holder, int amperage) {
+    protected static long add(long holder, long amperage) {
         return (long) getMaxAmperage(holder) << 32 | getAmperage(holder) + amperage;
     }
